feat: 게시판(Board)이나 사용자(Member) 삭제 시 해당 게시판에 작성된 게시물이 하나라도 있으면 400을 반환하는 예외 처리 추가

+++ src/main/java/com/example/bcsd/service/ArticleService.java
@@ -12,7 +12,6 @@ import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
 import java.util.List;
-import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -104,6 +103,7 @@ public class ArticleService {
     }
 
     public void deleteArticle(Long id) {
+        if (articleDao.isExistArticleByAuthorId(id))
         articleDao.deleteArticle(id);
     }
